trying to get legacy android <12 version to work with bluetooth, but going to give up on that for now, and focus on android version 12+ instead.
